# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A small in-memory CRUD API for managing tasks. Built for FlyRank Backend AI Engineering — BE-01."
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "responses": { "200": { "description": "Basic info about this API" } }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": { "200": { "description": "Server is alive" } }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "parameters": [
          { "name": "done", "in": "query", "schema": { "type": "boolean" }, "description": "Filter by completion status" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Filter by title substring" }
        ],
        "responses": { "200": { "description": "Array of tasks" } }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "title": { "type": "string", "example": "Buy milk" } },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Missing or empty title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get one task by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "The task" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated task" },
          "400": { "description": "Invalid body" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "Deleted, no content" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Task counts",
        "responses": { "200": { "description": "total/done/open counts" } }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset to the 3 seed tasks",
        "responses": { "200": { "description": "Tasks reset" } }
      }
    }
  }
}